Move types to their own folder
