+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.IoTHub/stream/IStreamRequestHandler.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Agent.IoTHub.Stream
+{
+    using System.Net.WebSockets;
+    using System.Threading;
+    using System.Threading.Tasks;
+
+    public interface IStreamRequestHandler
+    {
+        Task Handle(ClientWebSocket clientWebSocket, CancellationToken cancellationToken);
+    }
+}
