+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.IoTHub/stream/IStreamRequestListener.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Agent.IoTHub.Stream
+{
+    public interface IStreamRequestListener
+    {
+        void InitPump(IModuleClient moduleClient);
+    }
+}
